get withdraw fee, separate kraken providers and add asset to all providers
