+++ src/Provider/WithdrawAddressProviderInterface.php
@@ -7,7 +7,11 @@ namespace Jorijn\Bitcoin\Dca\Provider;
 interface WithdrawAddressProviderInterface
 {
     /**
-     * Method should return a Bitcoin address for withdrawal.
+     * Method should return the asset of the withdrawal asset.
+     */
+    public function getAsset(): string;
+    /**
+     * Method should return a address for withdrawal.
      */
     public function provide(): string;
 }
